Refine backtracking behavior
To avoid that we backtrack after parsing arbitrarily deep into branches
of alternatives we have to manually specify the intended amount of lookahead.

We implement this by *not backtracking* on `p` in `p or q` whenever
`p` consumed any input. The caller then wraps the parser for some prefix
of `q` that commits the branch in `tryOrRewind`. If parsing fails in
that prefix, then no input is consumed and we backtrack. If parsing fails
in `p` after parsing the rewindable prefix then input is consuemd and we do
not backtrack but fail `p or q`.

This strategy follows what Parsec and Megaparsec do.